% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACHESIS.R
\name{.ecdf.stats}
\alias{.ecdf.stats}
\title{Compute ECDF statistics with lower and upper bounds}
\usage{
.ecdf.stats(data, mean, lower, upper)
}
\arguments{
\item{data}{data table containing mean, lower and upper bound of value of interest.}

\item{mean}{column name containing mean values.}

\item{lower}{column name containing lower bounds.}

\item{upper}{column name containing upper bounds.}
}
\description{
Compute ECDF statistics with lower and upper bounds
}
