% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plot.CNV}
\alias{plot.CNV}
\title{Plot a tumor's copy number profile.}
\usage{
plot.CNV(cn.data)
}
\arguments{
\item{cn.data}{A data frame with copy number information as produced by readCNV.R}
}
\value{
A standardized data frame with copy number information per segment.
plot.CNV()
}
\description{
Plots the copy numbers used for mutation density information.
}
\examples{
aceseq.cn = system.file("extdata", "ACESeq/NBE11_comb_pro_extra2.59_0.83.txt", package = "NBevolution")
cn.data = readCNV(aceseq.cn)
plot.CNV(cn.data)
}
\author{
{Verena Körber}
}
