% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACHESIS.R
\name{plotClinicalCorrelations}
\alias{plotClinicalCorrelations}
\title{Correlate SNV density at ECA/MRCA with clinical parameters such as age, OS, etc.}
\usage{
plotClinicalCorrelations(
  lachesis = NULL,
  clin.par = "Age",
  suppress.outliers = FALSE,
  log.densities = FALSE,
  output.file = NULL
)
}
\arguments{
\item{lachesis}{output generated from \code{\link{LACHESIS}}}

\item{clin.par}{the clinical parameter used for correlation. Default \code{Age}.}

\item{suppress.outliers}{shall outliers (defined as the 2.5\% tumors with lowest and highest densities) be plotted? Default \code{TRUE}.}

\item{log.densities}{plot logarithmic densities. Default \code{FALSE}}

\item{output.file}{optional; the file to which the plot will be stored.}
}
\description{
Takes SNV densities as computed by \code{LACHESIS} as input and correlates them with clinical data such as age at diagnosis, survival data etc.
}
\examples{
input.files = system.file("extdata", "Sample_template.txt")
lachesis <- LACHESIS(input.files = input.files)
plotClinicalCorrelations(lachesis)
}
