% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_NB_initiation.R
\name{simulateCI}
\alias{simulateCI}
\title{Simulate mutation densities and their 95\% CI using population genetic model of neuroblastoma initiation.}
\usage{
simulateCI(
  parameter.samples,
  measured.mutation.times,
  measured.mutation.times.lr = NA,
  measured.mutation.times.eca,
  mode = "expansion_only"
)
}
\arguments{
\item{parameter.samples}{a data frame with the parameter samples}

\item{mode}{Whether tumor initiation should be modeled in a homeostatic tissue ("homeostasis") or in a transient tissue ("decay")}

\item{P.MRCA}{a data frame with the measured densities at MRCA among high-risk tumors (TMM). Rownames must identify the individual tumors. Must contain 1 column named density in units #SNVs/haploid genome}

\item{P.MRCA.lr}{a data frame with the measured densities at MRCA among low-risk tumors (no TMM). Rownames must identify the individual tumors. Must contain 1 column named density in units #SNVs/haploid genome}

\item{P.ECA}{a data frame with the measured densities at ECA among high-risk tumors (TMM). Rownames must identify the individual tumors. Must contain 1 column named density in units #SNVs/haploid genome}
}
\value{
A list containing lower and upper bounds of the 95\% CI of the simulated probabilities at the measured incidences of ECA and MRCA
simulateCI()
}
\description{
This method takes as input the parameter estimates obtained by fitting a population genetics model to NB initiation. It then simulates the model using the parameter estimates, which can be useful for plotting the fitting result.
}
