% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMutationRate.R
\name{estimateMutationRate}
\alias{estimateMutationRate}
\title{Estimate the mutation rate from a cohort of tumors with age information}
\usage{
estimateMutationRate(lachesis = NULL, output.dir = NULL, overwrite = FALSE)
}
\arguments{
\item{lachesis}{output generated from \code{\link{LACHESIS}}}

\item{overwrite}{logical; if \code{FALSE}, the function will not be run again if
the mutation rate has already been estimated. Instead, the available estimate
will be read in.}

\item{...}{further arguments and parameters passed to other}

\item{output.file}{optional; will save the plot.}
}
\value{
mean, standard deviation and 95\% confidence interval for the tumor
growth time and the mutation rate (per Mb), along with goodness of fit
(R squared).
}
\description{
This function fits a linear model to infer the mutation rate from the
relationship between age at diagnosis and mutational burden at MRCA. The
model assumes a constant mutation rate prior to the MRCA and comparable
tumor growth rates between individuals. Accordingly, the mutational burden
per Mb at MRCA is Poisson distributed around $$\mu t_\mathrm{MRCA}$$, where
$$\mu$$ is the mutation rate per Mb and $t_\mathrm{MRCA}$$ is the time at
which the MRCA emerges.
}
\examples{
# An example file with sample annotations and meta data
input.files <- system.file("extdata", "Sample_template.txt",
    package = "LACHESIS"
)
input.files <- data.table::fread(input.files)

# cnv and snv files for example tumors
nbe11 <- list.files(system.file("extdata/NBE11/", package = "LACHESIS"),
    full.names = TRUE
)
nbe15 <- list.files(system.file("extdata/NBE15/", package = "LACHESIS"),
    full.names = TRUE
)
nbe26 <- list.files(system.file("extdata/NBE26/", package = "LACHESIS"),
    full.names = TRUE
)

cnv.file <- c(nbe11[1], nbe15[1], nbe26[1])
snv.file <- c(nbe11[2], nbe15[2], nbe26[2])

input.files$cnv.file <- cnv.file
input.files$snv.file <- snv.file

# Make an example input file with paths to cnv and snv file along with other
# meta data
lachesis_input <- tempfile(
    pattern = "lachesis", tmpdir = tempdir(),
    fileext = ".tsv"
)
data.table::fwrite(x = input.files, file = lachesis_input, sep = "\t")

# Example with template file with paths to multiple cnv/snv files as an input
lachesis <- LACHESIS(input.files = lachesis_input)
estimateMutationRate(lachesis)

}
