% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACHESIS.R
\name{LACHESIS}
\alias{LACHESIS}
\title{Run MRCA density estimation for a set of tumors}
\usage{
LACHESIS(
  input.files = NULL,
  ids = NULL,
  cnv.files = NULL,
  snv.files = NULL,
  vcf.source = NULL,
  purity = NULL,
  ploidy = NULL,
  cnv.chr.col = NULL,
  cnv.start.col = NULL,
  cnv.end.col = NULL,
  cnv.A.col = NULL,
  cnv.B.col = NULL,
  cnv.tcn.col = NULL,
  age = NULL,
  OS.time = NULL,
  OS = NULL,
  EFS.time = NULL,
  EFS = NULL,
  output.dir = NULL,
  ...
)
}
\arguments{
\item{input.files}{a tab-delimited sample-specification file, it must contain the sample name, the path to the SNV file, path to CNV file, and optionally purity, ploidy, cnv.chr.col, cnv.start.col, cnv.end.col, cnv.A.col, cnv.B.col, cnv.tcn.col. A template for this spreadsheet can be downloaded from ...}

\item{ids}{vector of sample names, will be ignored if \code{input.files} is specified.}

\item{cnv.files}{vector of cnv files in same order as ids; should be in tab-delimited format, will be ignored if \code{input.files} is specified.}

\item{snv.files}{vector of snv files in same order as ids; should be in vcf format, will be ignored if \code{input.files} is specified.}

\item{vcf.source}{Tool used for generating VCF file. Can be \code{strelka} or \code{mutect} or \code{dkfz}}

\item{purity}{vector tumor cell content in same order as ids; will be ignored if \code{input.files} is specified.}

\item{ploidy}{average copy number in the tumor sample in same order as ids; will be ignored if \code{input.files} is specified.}

\item{cnv.chr.col}{column index of chromosome number in cnv.files}

\item{cnv.start.col}{column index of first position of the segment}

\item{cnv.end.col}{column index of last position of the segment}

\item{cnv.A.col}{column index of the number of A alleles. If A and B are not provided, allele configuration are assumed as 1:1 for disomic, 2:1 for trisomic and 3:1 for tetrasomic regions.}

\item{cnv.B.col}{column index of the number of B alleles. If A and B are not provided, allele configuration are assumed as 1:1 for disomic, 2:1 for trisomic and 3:1 for tetrasomic regions.}

\item{cnv.tcn.col}{column index of the total copy number. Is computed to A + B if not provided.}

\item{age, }{optional, the age at diagnosis}

\item{OS.time, }{optional, overall survival time}

\item{OS, }{optional, overall survival indicator variable}

\item{EFS.time, }{optional, event-free survival time}

\item{EFS, }{optional, event-free survival indicator variable
optional arguments}

\item{output.dir}{link to directory in which output is to be stored.}

\item{min.vaf}{Remove variants with vcf below threshold. Default 0.01}

\item{min.depth}{Minimum required depth for a variant to be considered. Default 30.}

\item{min.cn}{maximum copy number to be included in the plotting. Defaults to 2.}

\item{max.cn}{maximum copy number to be included in the analysis. Defaults to 4.}

\item{chromosomes}{vector of chromosomes to include in SNV density estimation}

\item{merge.tolerance}{the maximum distance below which adjacent segments with equal copy number are merged. Defaults to 10^5 bp.}

\item{ignore.XY}{Ignore allosomes. Default TRUE}

\item{ref_build}{Reference genome. Default \code{hg19}. Can be \code{hg18}, \code{hg19} or \code{hg38}}
}
\value{
a data.table
}
\description{
Takes a set of SNV and CNV files as input and outputs per-tumor SNV densities. Input can either be a tab-delimited file containing
the sample specifications or vectors giving direct paths to the sample files. CNV file requires columns for the chromosome number, start and end of the segment, and either the total copy number or the number of A- and B-alleles
}
\examples{
#an example file with sample annotations and meta data
input.files = system.file("extdata", "Sample_template.txt", package = "LACHESIS")
input.files = data.table::fread(input.files)

#cnv and snv files for example tumors
nbe11 = list.files(system.file("extdata/NBE11/", package = "LACHESIS"), full.names = TRUE)
nbe15 = list.files(system.file("extdata/NBE15/", package = "LACHESIS"), full.names = TRUE)
nbe63 = list.files(system.file("extdata/NBE63/", package = "LACHESIS"), full.names = TRUE)

cnv.file = c(nbe11[1], nbe15[1], nbe63[1])
snv.file = c(nbe11[2], nbe15[2], nbe63[2])

input.files$cnv.file = cnv.file
input.files$snv.file = snv.file

# Make an axample input file with paths to cnv and snv file along with other meta data
lachesis_input = tempfile(pattern = "lachesis", tmpdir = tempdir(), fileext = ".tsv")
data.table::fwrite(x = input.files, file = lachesis_input, sep = "\t")

#Exampele with template file with paths to multiple cnv/snv files as an input
lachesis <- LACHESIS(input.files = lachesis_input)

#Example with a single sample input
strelka_vcf = system.file("extdata","strelka2.somatic.snvs.vcf.gz", package = "LACHESIS")
aceseq_cn = system.file("extdata", "ACESeq/NBE11_comb_pro_extra2.59_0.83.txt", package = "LACHESIS")
lachesis <- LACHESIS(ids = "NBE11", cnv.files = aceseq_cn, snv.files = strelka_vcf, vcf.source = "strelka", purity = 0.83, ploidy = 2.59)
}
\seealso{
\code{\link{MRCA}} \code{\link{clonalMutationCounter}} \code{\link{normalizeCounts}}
}
