% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbImport.R
\name{nbImport}
\alias{nbImport}
\title{Combine CNVs and SNVs}
\usage{
nbImport(cnv = NULL, snv = NULL)
}
\arguments{
\item{cnv}{CNV data from \code{\link{readCNV}}}

\item{snv}{SNV data from \code{\link{readVCF}}}
}
\value{
a data.table
}
\description{
Merges CNVs and SNVs into a single data.table. Each variant is assigned to its corresponding copy number segment and status.
}
\examples{
snvs <- system.file("extdata", "snvs_NBE15_somatic_snvs_conf_8_to_10.vcf", package = "NBevolution")
s_data <- readVCF(vcf = snvs, vcf.source = "dkfz")
aceseq_cn <- system.file("extdata", "NBE15_comb_pro_extra2.51_1.txt", package = "NBevolution")
c_data <- readCNV(aceseq_cn)
nb <- nbImport(cnv = c_data, snv = s_data)
}
\seealso{
\code{\link{plotNB}}
}
