% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbImport.R
\name{nbImport}
\alias{nbImport}
\title{Combine CNVs and SNVs}
\usage{
nbImport(cnv, snv)
}
\arguments{
\item{cnv}{CNV data from \code{\link{readCNV}}}

\item{snv}{SNV data from \code{\link{readVCF}}}
}
\value{
a data.table
}
\description{
Merges CNVs and SNVs into a single data.frame. Each variant is assigned to its corresponding copy number segment and status.
}
\examples{
snvs <- system.file("extdata", "NBE15/NBE15.snvs.dkfz.tsv.gz", package = "NBevolution")
s_data <- readVCF(vcf = snvs, vcf.source = "dkfz")
aceseq_cn <- system.file("extdata", "NBE15/NBE15.aceseq.tsv.gz", package = "NBevolution")
c_data <- readCNV(aceseq_cn)
nb <- nbImport(cnv = c_data, snv = s_data)
}
