% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiseaseTrajectories.R
\name{plotDiseaseTrajectories}
\alias{plotDiseaseTrajectories}
\title{Plot the disease trajectories in a cohort.}
\usage{
plotDiseaseTrajectories(
  lachesis = NULL,
  mut.snv.rate = 3.2,
  estimate.mut.rate = FALSE,
  lach.col.eca = "#176A02",
  lach.col.mrca = "#4FB12B",
  corr.time.scale = 1,
  output.file = NULL
)
}
\arguments{
\item{lachesis}{output generated from \code{\link{LACHESIS}}.}

\item{mut.snv.rate}{optional; rate of accumulated SNVs per day in a
diploid genome (e.g. 3.2 SNVs/day in neuroblastoma). Will be ignored if
\code{estimate.mut.rate} is set to \code{TRUE}}

\item{estimate.mut.rate}{if set to \code{TRUE}, the mutation rate will be
estimated by fitting a linear regression model to the relationship between
age at diagnosis and mutaiton density at MRCA. Default \code{FALSE}.}

\item{lach.col.eca}{optional, color for ECA.}

\item{lach.col.mrca}{optional, color for MRCA.}

\item{corr.time.scale}{numeric value by which survival time is to be divided
to convert into months (e.g., 30 for converting days into months), defaults
to \code{1}.}

\item{output.file}{optional; file path to output.}
}
\value{
graph with SNV densities and estimated times at ECA/ MRCA and
diagnosis.
}
\description{
Takes SNV densities as computed by \code{LACHESIS} as input and plots for each
case the estimated age at ECA, MRCA and the age at diagnosis.
}
\examples{
# An example file with sample annotations and meta data
input.files <- system.file("extdata", "Sample_template.txt",
    package = "LACHESIS"
)
input.files <- data.table::fread(input.files)

# cnv and snv files for example tumors
nbe11 <- list.files(system.file("extdata/NBE11/", package = "LACHESIS"),
    full.names = TRUE
)
nbe15 <- list.files(system.file("extdata/NBE15/", package = "LACHESIS"),
    full.names = TRUE
)
nbe26 <- list.files(system.file("extdata/NBE26/", package = "LACHESIS"),
    full.names = TRUE
)

cnv.file <- c(nbe11[1], nbe15[1], nbe26[1])
snv.file <- c(nbe11[2], nbe15[2], nbe26[2])

input.files$cnv.file <- cnv.file
input.files$snv.file <- snv.file

# Make an example input file with paths to cnv and snv file along with other
# meta data
lachesis_input <- tempfile(
    pattern = "lachesis", tmpdir = tempdir(),
    fileext = ".tsv"
)
data.table::fwrite(x = input.files, file = lachesis_input, sep = "\t")

# Example with template file with paths to multiple cnv/snv files as an input
lachesis <- LACHESIS(input.files = lachesis_input)
plotDiseaseTrajectories(lachesis, corr.time.scale = 31)
}
