% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACHESIS.R
\name{plotLachesis}
\alias{plotLachesis}
\title{Plot SNV densities at ECA and MRCA}
\usage{
plotLachesis(
  lachesis = NULL,
  suppress.outliers = FALSE,
  log.densities = FALSE,
  fill.multi = NULL,
  l.col = NULL,
  binwidth = NULL
)
}
\arguments{
\item{lachesis}{output generated from \code{\link{LACHESIS}}}

\item{suppress.outliers}{shall outliers (defined as the 2.5\% tumors with lowest and highest densities) be plotted? Default \code{TRUE}.}

\item{log.densities}{plot logarithmic densities. Default \code{FALSE}}

\item{fill.multi}{optional, the color code for ECA and MRCA.}

\item{l.col, }{optional, the line color}

\item{binwidth}{optional; the bin-width in the histogram.}
}
\description{
Visualizes results from \code{\link{LACHESIS}}. Top plot, histograms of mean mutation densities; optional, correlation with age at diagnosis; bottom plots, cumulative distribution of mean mutation densities with 95\% confidence intervals.
}
\examples{
input.files = system.file("extdata", "Sample_template.txt")
lachesis <- LACHESIS(input.files = input.files)
plotLachesis(lachesis)
}
