% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACHESIS.R
\name{plotLachesis}
\alias{plotLachesis}
\title{Plot SNV densities at ECA and MRCA}
\usage{
plotLachesis(
  lachesis = NULL,
  suppress.outliers = FALSE,
  log.densities = FALSE,
  fill.multi = NULL,
  l.col = NULL,
  binwidth = NULL,
  fill.zero = NULL,
  output.file = NULL
)
}
\arguments{
\item{lachesis}{output generated from \code{\link{LACHESIS}}}

\item{suppress.outliers}{shall outliers (defined as the 2.5\% tumors with lowest and highest densities) be plotted? Default \code{TRUE}.}

\item{log.densities}{plot logarithmic densities. Default \code{FALSE}}

\item{fill.multi}{optional, the color code for ECA.}

\item{l.col, }{optional, the line color}

\item{binwidth}{optional; the bin-width in the histogram.}

\item{fill.zero}{optional, the color code for MRCA.}

\item{output.file}{optional; the file to which the plot will be stored.}
}
\description{
Visualizes results from \code{\link{LACHESIS}}. Top plot, histograms of mean mutation densities; bottom plots, cumulative distribution of mean mutation densities with 95\% confidence intervals.
}
\examples{
#an example file with sample annotations and meta data
input.files = system.file("extdata", "Sample_template.txt", package = "LACHESIS")
input.files = data.table::fread(input.files)

#cnv and snv files for example tumors
nbe11 = list.files(system.file("extdata/NBE11/", package = "LACHESIS"), full.names = TRUE)
nbe15 = list.files(system.file("extdata/NBE15/", package = "LACHESIS"), full.names = TRUE)
nbe63 = list.files(system.file("extdata/NBE63/", package = "LACHESIS"), full.names = TRUE)

cnv.file = c(nbe11[1], nbe15[1], nbe63[1])
snv.file = c(nbe11[2], nbe15[2], nbe63[2])

input.files$cnv.file = cnv.file
input.files$snv.file = snv.file

# Make an example input file with paths to cnv and snv file along with other meta data
lachesis_input = tempfile(pattern = "lachesis", tmpdir = tempdir(), fileext = ".tsv")
data.table::fwrite(x = input.files, file = lachesis_input, sep = "\t")

#Example with template file with paths to multiple cnv/snv files as an input
lachesis <- LACHESIS(input.files = lachesis_input)
plotLachesis(lachesis)
}
