% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LACHESIS.R
\name{plotLachesis}
\alias{plotLachesis}
\title{Plot SNV densities at ECA and MRCA}
\usage{
plotLachesis(
  lachesis = NULL,
  suppress.outliers = FALSE,
  log.densities = FALSE,
  fill.multi = NULL,
  l.col = NULL,
  binwidth = NULL
)
}
\arguments{
\item{lachesis}{output generated from \code{\link{LACHESIS}}}

\item{suppress.outliers}{shall outliers (defined as the 2.5\% tumors with lowest and highest densities) be plotted? Default \code{TRUE}.}

\item{log.densities}{plot logarithmic densities. Default \code{FALSE}}

\item{fill.multi}{optional, the color code for ECA and MRCA.}

\item{l.col, }{optional, the line color}

\item{binwidth}{optional; the bin-width in the histogram.}
}
\description{
Visualizes results from \code{\link{LACHESIS}}. Top plot, histograms of mean mutation densities; bottom plots, cumulative distribution of mean mutation densities with 95\% confidence intervals.
}
\examples{
snvs <- system.file("extdata", "NBE15", "snvs_NBE15_somatic_snvs_conf_8_to_10.vcf", package = "LACHESIS")
s_data <- readVCF(vcf = snvs, vcf.source = "dkfz")
aceseq_cn <- system.file("extdata", "NBE15", "NBE15_comb_pro_extra2.51_1.txt", package = "LACHESIS")
c_data <- readCNV(aceseq_cn)
nb <- nbImport(cnv = c_data, snv = s_data, purity = 1, ploidy = 2.51)
cl_muts <- clonalMutationCounter(nb)
norm_muts <- normalizeCounts(cl_muts)
mrca <- MRCA(norm_muts)
plotMutationDensities(mrca)
}
