% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutationDensities.R
\name{plotMutationDensities}
\alias{plotMutationDensities}
\title{Plot normalized mutation density at copy number gain and MRCA per segment}
\usage{
plotMutationDensities(
  mrcaObj = NULL,
  samp.name = NULL,
  min.seg.size = 10^7,
  fill.zero = NULL,
  fill.multi = NULL,
  l.col = NULL,
  show.den = NULL,
  binwidth = NULL
)
}
\arguments{
\item{mrcaObj}{output generated from \code{\link{MRCA}}}

\item{samp.name}{sample name, optional}

\item{min.seg.size}{minimal segment size to plot}

\item{fill.zero}{optional, the color code for densities of mutations present on single copies.}

\item{fill.multi}{optional, the color code for densities of mutations present on multiple copies.}

\item{l.col, }{optional, the line color}

\item{show.den}{optional; if \code{TRUE}, the density distribution of mutation densities on single copies is shown in the histogram of mutation densities on multiple copies.}

\item{binwidth}{optional; the bin-width in the histogram.}
}
\description{
Visualizes results from  \code{\link{MRCA}}. Top plot, histograms of mean mutation densities; bottom plots, timeline of early tumor evolution, showing mutation densities (mean and 95\% CI) of individual chromosomal gains and mutation densities at ECA and MRCA.
}
\examples{
snvs <- system.file("extdata", "snvs_NBE15_somatic_snvs_conf_8_to_10.vcf", package = "NBevolution")
s_data <- readVCF(vcf = snvs, vcf.source = "dkfz")
aceseq_cn <- system.file("extdata", "NBE15_comb_pro_extra2.51_1.txt", package = "NBevolution")
c_data <- readCNV(aceseq_cn)
nb <- nbImport(cnv = c_data, snv = s_data, purity = 1, ploidy = 2.51)
cl_muts <- clonalMutationCounter(nb)
norm_muts <- normalizeCounts(cl_muts)
mrca <- MRCA(norm_muts)
plotMutationDensities(mrca)
}
