% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbImport.R
\name{plotNB}
\alias{plotNB}
\title{Plot VAF distribution per copy number}
\usage{
plotNB(
  nb = NULL,
  ref_build = "hg19",
  min.cn = 2,
  max.cn = 4,
  samp.name = NULL,
  purity = NULL,
  ploidy = NULL,
  output.file = NULL
)
}
\arguments{
\item{nb}{output generated from \code{\link{nbImport}}}

\item{ref_build}{Reference genome. Default \code{hg19}. Can be \code{hg18}, \code{hg19} or \code{hg38}}

\item{min.cn}{maximum copy number to be included in the plotting. Defaults to 2.}

\item{max.cn}{maximum copy number to be included in the plotting. Defaults to 4.}

\item{samp.name}{Sample name. Optional. Default NULL}

\item{purity}{optional. If \code{purity} and \code{ploidy} are both given, expected positions of clonal peaks are shown.}

\item{ploidy}{optional. If \code{purity} and \code{ploidy} are both given, expected positions of clonal peaks are shown.}

\item{output.file}{optional, will save the plot.}
}
\description{
Visualizes results from  \code{\link{nbImport}}. Top plot, measured copy numbers along the genome; bottom plots, VAF histograms of SNVs stratified by copy number.
}
\examples{
snvs <- system.file("extdata", "snvs_NBE15_somatic_snvs_conf_8_to_10.vcf", package = "LACHESIS")
s_data <- readVCF(vcf = snvs, vcf.source = "dkfz")
aceseq_cn <- system.file("extdata", "NBE15_comb_pro_extra2.51_1.txt", package = "LACHESIS")
c_data <- readCNV(aceseq_cn)
nb <- nbImport(cnv = c_data, snv = s_data)
plotNB(nb)
}
