% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbImport.R
\name{plotNB}
\alias{plotNB}
\title{Plot VAF distribution per copy number}
\usage{
plotNB(nb = NULL, ref_build = "hg19")
}
\arguments{
\item{nb}{output generated from \code{\link{nbImport}}}

\item{ref_build}{Default \code{hg19}. Can be \code{hg18}, \code{hg19} or \code{hg38}}
}
\description{
Visualize results from  \code{\link{nbImport}}
}
\examples{
snvs <- system.file("extdata", "NBE15/NBE15.snvs.dkfz.tsv.gz", package = "NBevolution")
s_data <- readVCF(vcf = snvs, vcf.source = "dkfz")
aceseq_cn <- system.file("extdata", "NBE15/NBE15.aceseq.tsv.gz", package = "NBevolution")
c_data <- readCNV(aceseq_cn)
nb <- nbImport(cnv = c_data, snv = s_data)
plotNB(nb)
}
