% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVAF.R
\name{plotVAFdistr}
\alias{plotVAFdistr}
\title{Plot histogram of VAF distribution}
\usage{
plotVAFdistr(vaf = NULL, vafbreak = 0.05, t_sample = NULL)
}
\arguments{
\item{vaf}{output produced by \code{\link{readVCF}}}

\item{vafbreak}{Interval size. Default 0.05}

\item{t_sample}{Sample name for tumor. Used for plot title. Default NULL}
}
\value{
data.table of frequency table
}
\description{
Plot frequency distribution of variant allele frequencies
}
\examples{
strelka_vcf = system.file("extdata", "strelka2.somatic.snvs.vcf.gz", package = "NBevolution")
s_data = readVCF(vcf = strelka_vcf, vcf_source = "strelka", ignore_XY = FALSE)
plotVAFdistr(s_data)
}
