% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCNV.R
\name{readCNV}
\alias{readCNV}
\title{Converts a user-specified bed-file with copy number information into a standardized format that can be used as input for downstream analysis}
\usage{
readCNV(
  cn.info,
  chr.col,
  start.col,
  end.col,
  A.col,
  B.col,
  merge.tolerance = 10^5
)
}
\arguments{
\item{cn.info}{A data frame containing the copy number information. Requires columns for the chromosome number, start and end of the segment, the number of A- and B-alleles}

\item{chr.col}{name or number of the column containing the chromosome number}

\item{start.col}{name or number of the column containing the first position of the segment}

\item{end.col}{name or number of the column containing the last position of the segment}

\item{A.col}{name or number of the column containing the number of A alleles}

\item{B.col}{name or number of the column containing the number of B alleles}

\item{merge.tolerance}{the maximal distance below which adjacent segments with equal copy number are merged. Defaults to 10^5 bp}
}
\value{
A standardized data frame with copy number information per segment.
readCNV()
}
\description{
Convert a user-specified bed-file with copy number information into a standardized format. Perform various quality checks on the file input and return the clean and standardized data-frame. Written on May 31st, 2023.
}
\author{
{Verena Körber}
}
