% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\title{Import a VCF file and extract read counts}
\usage{
readVCF(
  vcf = NULL,
  ignore_XY = TRUE,
  vcf_source = "strelka",
  min_vaf = 0.01,
  min_depth = 30,
  t_sample = NULL
)
}
\arguments{
\item{vcf}{Input indexed VCF file.}

\item{vcf_source}{Tool used for generating VCF file. Can be \code{strelka} ot \code{mutect}}

\item{min_vaf}{Remove variants with vcf below threshold. Default 0.01}

\item{min_depth}{Minimum required depth for a variant to be considered. Default 30.}

\item{t_sample}{Sample name for tumor. Must be same as in VCF. Strelka hardcodes tumor sample name to "TUMOR"}

\item{ignore_xy}{Ignore allosomes. Default TRUE}
}
\value{
a data.table with chrom, pos, ref, alt, t_depth, t_vaf
}
\description{
Import a VCF file and extract read-count and variant allele frequencies. Currently VCF files generated by \code{mutect2} and \code{strelka2} are supported.
}
\examples{
mutect_vcf = system.file("extdata", "mutect.somatic.vcf.gz", package = "NBevolution")
m_data = readVCF(vcf = mutect_vcf, vcf_source = "mutect")
strelka_vcf = system.file("extdata", "strelka2.somatic.snvs.vcf.gz", package = "NBevolution")
m_data = readVCF(vcf = strelka_vcf, vcf_source = "strelka")
}
