% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\title{Import a VCF file and extract read count}
\usage{
readVCF(
  vcf = NULL,
  ignore.XY = TRUE,
  vcf.source = "strelka",
  min.vaf = 0.01,
  min.depth = 30,
  t.sample = NULL
)
}
\arguments{
\item{vcf}{Input indexed VCF file.}

\item{ignore.XY}{Tool used for generating VCF file. Can be \code{strelka} ot \code{mutect}}

\item{min.vaf}{Remove variants with vcf below threshold. Default 0.01}

\item{min.depth}{Minimum required depth for a variant to be considered. Default 30.}

\item{t.sample}{Sample name for tumor. Must be same as in VCF. Strelka hardcodes tumor sample name to "TUMOR"}
}
\value{
a data.table with chrom, pos, ref, alt, t_ref_count, t_alt_count, t_depth, t_vaf
}
\description{
Import a VCF file and extract read-count and variant allele frequencies. Currently VCF files generated by \code{mutect2} and \code{strelka2} are supported.
}
\examples{
mutect_vcf = system.file("extdata", "mutect.somatic.vcf.gz", package = "NBevolution")
m_data = readVCF(vcf = mutect_vcf, vcf.source = "mutect")
strelka_vcf = system.file("extdata", "strelka2.somatic.snvs.vcf.gz", package = "NBevolution")
s_data = readVCF(vcf = strelka_vcf, vcf.source = "strelka")
}
